fix: compile warning add save mesh test add create box test add create cylinder test forceinline floathelper

+++ LearnDirectX12/Component/Geometry/HalfEdgeMesh.cpp
@@ -138,7 +138,7 @@ HEVertex *HEMesh::insertVertex(const float3 &position, const float2 &texcoord) {
 	auto *ptr = verts.back().get();
 	ptr->position = position;
 	ptr->texcoord = texcoord;
-	ptr->index = verts.size() - 1u;
+	ptr->index = static_cast<uint32>(verts.size()) - 1u;
 	return ptr;
 }
